Remove service package and refactor
