+++ pkg/node/node.go
@@ -38,6 +38,7 @@ import (
 	"github.com/ethersphere/bee/pkg/pusher"
 	"github.com/ethersphere/bee/pkg/pushsync"
 	"github.com/ethersphere/bee/pkg/recovery"
+	"github.com/ethersphere/bee/pkg/resolver"
 	resolverSvc "github.com/ethersphere/bee/pkg/resolver/service"
 	"github.com/ethersphere/bee/pkg/retrieval"
 	"github.com/ethersphere/bee/pkg/settlement/pseudosettle"
@@ -90,7 +91,7 @@ type Options struct {
 	GlobalPinningEnabled   bool
 	PaymentThreshold       uint64
 	PaymentTolerance       uint64
-	ResolverConnectionCfgs []*resolverSvc.ConnectionConfig
+	ResolverConnectionCfgs []*resolver.ConnectionConfig
 }
 
 func NewBee(addr string, logger logging.Logger, o Options) (*Bee, error) {
@@ -310,17 +311,7 @@ func NewBee(addr string, logger logging.Logger, o Options) (*Bee, error) {
 
 	b.pullerCloser = puller
 
-	// If any resolver endpoint is specified, init and connect the resolver.
-	if len(o.ResolverConnectionCfgs) > 0 {
-		logger.Debug("Connecting to name resolution services")
-
-		resolverSvc := resolverSvc.NewService(o.ResolverConnectionCfgs, logger)
-		resolverSvc.Connect()
-
-		b.resolverCloser = resolverSvc
-	} else {
-		logger.Info("No name resolution connection string specified")
-	}
+	b.resolverCloser = resolverSvc.InitMultiResolver(logger, o.ResolverConnectionCfgs)
 
 	var apiService api.Service
 	if o.APIAddr != "" {
